Added Specific App module named app_things, for Android Things.
Added a way to perform requests that are asynchronous and still module specific. The model is hidden to each module
Fixed GPS bug when GPS not present.
Added a state machine for the quality checker to avoid simultaenous operations such as compute and deinit
Device Name can now be set to null.
Device module has upgraded and uses now a new model of the MbtDevice object.